[BE][S13P31D104-119] refactor: 일부 엔티티 pk Short로 타입 변경

+++ backend/uniwiki/src/main/java/com/kiwi/uniwiki/domain/document/dto/request/DocumentUpdateRequestDTO.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 public class DocumentUpdateRequestDTO {
 
     @NotNull(message = "카테고리 선택은 필수입니다.")
-    private Integer categoryId;
+    private Short categoryId;
 
     private String documentContent;
